requirements.txt
djangorestframework
markdown
django
